update page title

+++ grails-app/services/stockfinancialinfo/StockFinancialInfoService.groovy
@@ -77,7 +77,7 @@ class StockFinancialInfoService {
                 stockFinancialInfoMap["noIndustryInfo"] = "暂无行业排名信息"
             }
             if (index == 'basicEPS') {
-                stockFinancialInfoMap["indexValue"] = ((int) stockFinancialInfo."${index}") + "元/每股"
+                stockFinancialInfoMap["indexValue"] = stockFinancialInfo."${index}" + "元/每股"
             } else {
                 stockFinancialInfoMap["indexValue"] = ((int) stockFinancialInfo."${index}") + "元"
             }
